passing score from python to C

+++ src/afl-fuzz.c
@@ -2277,7 +2277,7 @@ int main(int argc, char **argv_orig, char **envp) {
       msg_array[0] = afl->fsrv.map_size;
       
 
-      memcpy(send_data.data_buff, msg_array, BUFF_SIZE_SENDER * sizeof(double));
+      memcpy(send_data.data_buff, msg_array, BUFF_SIZE_SENDER * sizeof(u32));
       if (-1 == msgsnd(msqid_sender, &send_data, sizeof(t_send_u32_data) - sizeof(long), 0)) {
         perror("msgsnd() failed");
         exit(1);
